+++ internal/pg_documentdb_distributed/src/test/regress/sql/bson_aggregation_pipeline_tests_let.sql
@@ -334,13 +334,13 @@ RESET citus.max_adaptive_executor_pool_size;
 -- query match
 -- ignore variableSpec (turn both GUCs off)
 SET documentdb.enableLetAndCollationForQueryMatch TO off;
-SET documentdb.enableLetForWriteCommands TO off;
+SET documentdb.enableVariablesSupportForWriteCommands TO off;
 
 SELECT documentdb_api_internal.bson_query_match('{"a": "cat"}', '{"$expr": {"a": "$$varRef"} }', '{ "let": {"varRef": "cat"} }', NULL);
 
 -- use variableSpec (turn either GUC on)
 SET documentdb.enableLetAndCollationForQueryMatch TO on;
-SET documentdb.enableLetForWriteCommands TO on;
+SET documentdb.enableVariablesSupportForWriteCommands TO on;
 
 -- query match: wrong access of variable in query (outside $expr)
 SELECT documentdb_api_internal.bson_query_match('{"a": "cat"}', '{"a": "$$varRef"}', '{ "let": {"varRef": "cat"} }', NULL);
@@ -374,7 +374,7 @@ SELECT document from documentdb_api.collection('db', 'coll_query_op_let') WHERE 
 SELECT document from documentdb_api.collection('db', 'coll_query_op_let') WHERE  documentdb_api_internal.bson_query_match(document, '{"$expr": {"$not": {"$lt": ["$_id", "$$varRef"] } } }', '{ "let": {"varRef": 1} }', NULL);
 
 RESET documentdb.enableLetAndCollationForQueryMatch;
-RESET documentdb.enableLetForWriteCommands;
+RESET documentdb.enableVariablesSupportForWriteCommands;
 
 
 -- let with geonear