+++ internal/pg_documentdb_distributed/src/test/regress/sql/bson_dollar_ops_collation_tests_runtime.sql
@@ -761,13 +761,13 @@ SELECT document FROM bson_aggregation_pipeline('db', '{ "aggregate": "coll_agg_p
 -- query match
 -- ignore collation (turn both GUCs off)
 SET documentdb.enableLetAndCollationForQueryMatch TO off;
-SET documentdb.enableLetForWriteCommands TO off;
+SET documentdb.enableVariablesSupportForWriteCommands TO off;
 
 SELECT documentdb_api_internal.bson_query_match('{"a": "cat"}', '{"a": "CAT"}', NULL, 'en-u-ks-level1');
 
 -- enforce collation (turn either GUC on)
 SET documentdb.enableLetAndCollationForQueryMatch TO on;
-SET documentdb.enableLetForWriteCommands TO on;
+SET documentdb.enableVariablesSupportForWriteCommands TO on;
 
 -- query match: _id tests
 SELECT documentdb_api_internal.bson_query_match('{"_id": "cat"}', '{"_id": "CAT"}', NULL, 'en-u-ks-level1');
@@ -946,6 +946,6 @@ SELECT document FROM bson_aggregation_find('db', '{ "find": "nested_arrays_docs"
 SELECT document FROM bson_aggregation_find('db', '{ "find": "nested_arrays_docs", "filter": { "a" : {"$in" : [ {"b": [["dOg"]] } ] }}, "sort": { "_id": 1 }, "skip": 0, "limit": 5, "collation": { "locale": "en", "strength" : 1} }');
 
 RESET documentdb.enableLetAndCollationForQueryMatch;
-RESET documentdb.enableLetForWriteCommands;
+RESET documentdb.enableVariablesSupportForWriteCommands;
 
 RESET documentdb_core.enablecollation;